+++ buch/papers/neuronal/src/Residuals.py
@@ -1,0 +1,22 @@
+import torch
+
+# Define the wave equation residual
+# Equation 23.4 in the paper
+def wave_equation_residual(model, xyt, c=2.0):
+    xyt.requires_grad_(True)
+    u = model(xyt)
+
+    # first-order derivatives
+    grad_u = torch.autograd.grad(u, xyt, grad_outputs=torch.ones_like(u), create_graph=True)[0]
+    u_t = grad_u[:, 2:3]  # du/dt
+
+    # second-order derivatives
+    grad_u_xyt = torch.autograd.grad(u_t, xyt, grad_outputs=torch.ones_like(u_t), create_graph=True)[0]
+    u_tt = grad_u_xyt[:, 2:3]  # d^2u/dt^2
+
+    laplacian_u = \
+    torch.autograd.grad(grad_u[:, 0:2], xyt, grad_outputs=torch.ones_like(grad_u[:, 0:2]), create_graph=True)[0]
+    laplacian_u = laplacian_u[:, 0:1] + laplacian_u[:, 1:2]  # d^2u/dx^2 + d^2u/dy^2
+
+    residual = u_tt - c ** 2 * laplacian_u  # Equivalent to equation 23.4
+    return residual ** 2
